feat(gemini): unify antigravity and jetski integrations under adlc-gemini package
