+++ apps/docs/content/docs/integrations/index.mdx
@@ -1,6 +1,6 @@
 ---
 title: Integrations
-description: "Native ADLC integrations for six agentic coding harnesses: Claude Code, Codex, Cursor, opencode, pi, and Antigravity."
+description: "Native ADLC integrations for Google Gemini (Antigravity/JetSki), Claude Code, Codex, Cursor, opencode, and pi."
 ---
 
 # Integrations
@@ -20,7 +20,7 @@ harness below.
   lifecycle commands, a five-lens prosecutor panel, and a keyless gate bridge.
 - **[pi](/docs/integrations/pi)** is a native extension with proactive rail blocking,
   git-diff revert, and a suppression gate. Five phase skills wrap the CLIs.
-- **[Antigravity](/docs/integrations/antigravity)** is a native `agy` plugin with an
+- **[Google Gemini (Antigravity & JetSki)](/docs/integrations/gemini)** is a native plugin with an
   in-session rails-guard hook, backed by the unbypassable CI rail-freeze gate.
 
 ## Any other agent
@@ -49,7 +49,7 @@ These integrations wire the gates *into* each harness. The
 dispatches ready tickets to sandboxed **headless workers** in isolated worktrees
 and merges them through the gates. Its scheduler is harness-blind via a
 `WorkerAdapter` seam. Claude Code is the default worker, and adapters for Codex,
-Antigravity, opencode, pi, and Cursor let a single fleet run build tickets on any
+Antigravity/JetSki (Gemini), opencode, pi, and Cursor let a single fleet run build tickets on any
 of them (the harness is an operator-local `--adapter` choice).
 
 See the [toolkit overview](/docs/lifecycle) and the [ADLC theory](/docs/theory) for
