+++ apps/docs/content/docs/integrations/meta.json
@@ -1,1 +1,1 @@
-{ "title": "Integrations", "pages": ["index", "claude-code", "codex", "cursor", "opencode", "pi", "antigravity", "copilot"] }
+{ "title": "Integrations", "pages": ["index", "claude-code", "codex", "cursor", "opencode", "pi", "gemini", "copilot"] }
