+++ plugins/adlc-gemini/agents/prosecutor-contract.md
