+++ plugins/adlc-gemini/agents/prosecutor-diff.md
