+++ plugins/adlc-gemini/agents/prosecutor-security.md
