+++ plugins/adlc-gemini/agents/prosecutor-verifier.md
