+++ plugins/adlc-gemini/agents/prosecutor.md
