+++ plugins/adlc-gemini/core-inline.mjs
