+++ plugins/adlc-gemini/generated-active-ticket.mjs
