+++ plugins/adlc-gemini/generated-ticket-reader.mjs
