+++ plugins/adlc-gemini/hooks/adlc-rails-guard.cjs
