+++ plugins/adlc-gemini/plugin.json
@@ -1,0 +1,6 @@
+{
+  "name": "adlc-gemini",
+  "version": "1.7.0",
+  "adlcContract": 1,
+  "description": "Embrace the ADLC inside Google Gemini (antigravity/jetski): phase-router + doctrine skills and a PreToolUse rails-guard that freezes ADLC rails. Requires the @adlc/cli gate toolkit (npm i -g @adlc/cli)."
+}
