+++ plugins/adlc-gemini/rails-checker.mjs
