+++ plugins/adlc-gemini/skills/adlc-doctrine/SKILL.md
