+++ plugins/adlc-gemini/skills/adlc-prosecutor/SKILL.md
