+++ plugins/adlc-gemini/skills/adlc-self-orchestrate/SKILL.md
