+++ plugins/adlc-gemini/test/manifest.test.mjs
