+++ scripts/gemini-install-smoke.mjs
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
-// antigravity-install-smoke.mjs — verify the adlc-antigravity package shape,
+// gemini-install-smoke.mjs — verify the adlc-gemini package shape,
 // hooks.json schema (V3), always-exit-0 deny contract (V5), $HOME command path,
-// name/dir invariant, doc framing, and run the plugin unit tests. No agy binary
+// name/dir invariant, doc framing, and run the plugin unit tests. No agent binary
 // required. Exit 0 = pass, 2 = fail.
 import { execFileSync } from 'node:child_process';
 import { mkdirSync, mkdtempSync, readFileSync, readdirSync, writeFileSync } from 'node:fs';
@@ -10,15 +10,15 @@ import { dirname, join, resolve } from 'node:path';
 import { fileURLToPath } from 'node:url';
 
 const ROOT = resolve(process.argv[2] ?? join(dirname(fileURLToPath(import.meta.url)), '..'));
-const PLUGIN = join(ROOT, 'plugins', 'adlc-antigravity');
+const PLUGIN = join(ROOT, 'plugins', 'adlc-gemini');
 let failures = 0;
-const fail = (m) => { console.error(`antigravity-install-smoke: FAIL — ${m}`); failures++; };
+const fail = (m) => { console.error(`gemini-install-smoke: FAIL — ${m}`); failures++; };
 const ok = (m) => console.log(`  ok — ${m}`);
 const read = (p) => readFileSync(p, 'utf8');
 
 // manifest + name/dir invariant (F3)
 const manifest = JSON.parse(read(join(PLUGIN, 'plugin.json')));
-if (manifest.name !== 'adlc-antigravity') fail(`plugin.json name is ${manifest.name}`); else ok('plugin.json name == dir name');
+if (manifest.name !== 'adlc-gemini') fail(`plugin.json name is ${manifest.name}`); else ok('plugin.json name == dir name');
 
 // hooks.json V3 schema + $HOME command + catch-all matcher + .cjs entry
 const hj = JSON.parse(read(join(PLUGIN, 'hooks.json')));
@@ -26,7 +26,7 @@ const spec = hj['adlc-rails']?.PreToolUse?.[0];
 if (!spec) fail('hooks.json: adlc-rails.PreToolUse[0] missing'); else ok('hooks.json V3 shape (named hook → PreToolUse array)');
 if (spec?.matcher !== '.*') fail(`matcher is "${spec?.matcher}", not catch-all`); else ok('catch-all matcher');
 const cmd = spec?.hooks?.[0]?.command ?? '';
-if (!/\$HOME\/\.gemini\/config\/plugins\/adlc-antigravity\/hooks\/adlc-rails-guard\.cjs/.test(cmd)) fail(`command is not the $HOME .cjs path: ${cmd}`); else ok('command uses $HOME .cjs path (V9)');
+if (!/\$HOME\/\.gemini\/config\/plugins\/adlc-gemini\/hooks\/adlc-rails-guard\.cjs/.test(cmd)) fail(`command is not the $HOME .cjs path: ${cmd}`); else ok('command uses $HOME .cjs path (V9)');
 
 // deny path: only node: + @adlc/core
 const chk = read(join(PLUGIN, 'rails-checker.mjs'));
@@ -34,7 +34,7 @@ const guard = read(join(PLUGIN, 'hooks', 'adlc-rails-guard.mjs'));
 const imports = [...chk.matchAll(/from '([^']+)'/g), ...guard.matchAll(/from '([^']+)'/g)].map((m) => m[1]);
 if (imports.some((s) => !s.startsWith('node:') && !s.startsWith('.') && s !== '@adlc/core')) fail('deny path imports third-party deps'); else ok('deny path: node: + @adlc/core only');
 
-// self-contained portability check (Task-10 live-gate finding): `agy plugin install`
+// self-contained portability check: `[agent] plugin install`
 // COPIES this plugin into ~/.gemini/config/plugins/<name>/ WITHOUT node_modules, so
 // ANY runtime import of a workspace package (e.g. '@adlc/core') fails to resolve and
 // the hook fails closed on every tool. The plugin must be fully self-contained —
@@ -48,17 +48,14 @@ const SELF_CONTAINED_FILES = [
 ];
 let selfContained = true;
 for (const [label, src] of SELF_CONTAINED_FILES) {
-  // Match an actual import/require of the package, not prose in a comment (the
-  // header comments legitimately reference '@adlc/core' by name when explaining
-  // the history of this port).
   if (/from\s+['"]@adlc\/core['"]|require\(\s*['"]@adlc\/core['"]\s*\)/.test(src)) {
-    fail(`${label} imports '@adlc/core' — plugin must be self-contained — \`agy plugin install\` copies without node_modules`);
+    fail(`${label} imports '@adlc/core' — plugin must be self-contained — copies without node_modules`);
     selfContained = false;
   }
   for (const m of src.matchAll(/from '([^']+)'/g)) {
     const spec = m[1];
     if (!spec.startsWith('node:') && !spec.startsWith('.')) {
-      fail(`${label} imports non-node/non-relative module "${spec}" — plugin must be self-contained — \`agy plugin install\` copies without node_modules`);
+      fail(`${label} imports non-node/non-relative module "${spec}" — plugin must be self-contained — copies without node_modules`);
       selfContained = false;
     }
   }
@@ -66,7 +63,7 @@ for (const [label, src] of SELF_CONTAINED_FILES) {
 if (selfContained) ok('plugin is self-contained (node: + relative imports only, no @adlc/core)');
 
 // always exit 0 + allow_tool contract: drive the shim with a rail-hit fixture
-const repo = mkdtempSync(join(tmpdir(), 'agy-smoke-'));
+const repo = mkdtempSync(join(tmpdir(), 'gemini-smoke-'));
 mkdirSync(join(repo, '.adlc'), { recursive: true });
 writeFileSync(join(repo, '.adlc', 'tickets.json'), JSON.stringify({ tickets: [{ id: 'T1', title: 't', body: 'b', scope: ['src/**'], rails: ['src/frozen.js'] }] }));
 writeFileSync(join(repo, '.adlc', 'current-ticket.json'), JSON.stringify({ id: 'T1' }));
@@ -80,10 +77,10 @@ if (drive('write_to_file', { TargetFile: join(repo, 'src', 'frozen.js') }).allow
 if (drive('write_to_file', { TargetFile: join(repo, 'src', 'ok.js') }).allow_tool !== true) fail('shim did not ALLOW a non-rail write'); else ok('shim allows a non-rail write');
 
 // doc framing
-const doc = read(join(ROOT, 'docs', 'integrations', 'antigravity.md'));
+const doc = read(join(ROOT, 'docs', 'integrations', 'gemini.md'));
 if (!/rails-guard-ci\.mjs|ci\/rails-guard\.yml/.test(doc)) fail('doc does not name the CI gate'); else ok('doc names the CI gate');
 if (!/advisor/i.test(doc)) fail('doc does not frame the hook as advisory'); else ok('doc frames hook as advisory');
-if (!/Formal ADLC Coverage/.test(doc)) fail('doc missing Formal ADLC Coverage table'); else ok('doc has coverage table');
+if (!/Formal ADLC Coverage/i.test(doc)) fail('doc missing Formal ADLC Coverage table'); else ok('doc has coverage table');
 
 // run the plugin unit tests
 try {
@@ -92,5 +89,5 @@ try {
   ok('plugin unit tests pass');
 } catch (e) { fail(`unit tests failed:\n${e.stdout?.toString() ?? e.message}`); }
 
-if (failures) { console.error(`\nantigravity-install-smoke: ${failures} failure(s)`); process.exit(2); }
-console.log('\nantigravity-install-smoke: PASS');
+if (failures) { console.error(`\ngemini-install-smoke: ${failures} failure(s)`); process.exit(2); }
+console.log('\ngemini-install-smoke: PASS');
